refactor: extract autocomplete to its own file
